refactor(tests): make transaction sync/async tests more uniform
