+++ tests/test_transaction_async.py
@@ -6,9 +6,9 @@ import pytest
 from psycopg import Rollback
 from psycopg import errors as e
 
-from .test_transaction import in_transaction, insert_row, inserted, get_exc_info
-from .test_transaction import ExpectedException, crdb_skip_external_observer
-from .test_transaction import create_test_table  # noqa  # autouse fixture
+from ._test_transaction import in_transaction, insert_row, inserted, get_exc_info
+from ._test_transaction import ExpectedException, crdb_skip_external_observer
+from ._test_transaction import create_test_table  # noqa  # autouse fixture
 
 
 @pytest.fixture
